+++ pepper/rc/changelog.md
@@ -5,12 +5,13 @@
 - changed string syntax for commands
 - command strings now support some escapings
 - command aliases that start with `-` won't show up in auto completions
-- removed `find-file` and `find-command` commands as they're now implementable using other builtin commands (see `default_commands.pepper` for an example)
 - merged all `map-<mode>` commands into a single `map` command whose first parameter is the mode to map keys to
 - merged all `syntax-<token-kind>` commands into the `syntax` command which can take the first parameter the token kind for the defined pattern
 - insert processes now correctly adjust their insert positions on buffer insertions and deletions
 - added `set-register` command
 - changed `open` command parameters order, now buffer properties come before the `path` parameter
+- removed `alias` command since it's now possible to replicate its behavior by creating a new command that calls the aliased command and use the `@arg()` expansion
+- removed `find-file` and `find-command` commands as they're now implementable using other builtin commands (see `default_commands.pepper` for an example)
 
 ## 0.24.0
 - handle buffer paths beginning with `./` (on `Buffer::set_path` and `Buffer::find_with_path`)